move MatrixAutorizationInfo and MatrixUserInfo to rest package
