+++ src/main/java/org/beta/authentification/keycloak/bot/BotSender.java
@@ -5,7 +5,7 @@
 package org.beta.authentification.keycloak.bot;
 
 import org.beta.authentification.keycloak.utils.LoggingUtilsFactory;
-import org.beta.authentification.matrix.MatrixUserInfo;
+import org.beta.authentification.matrix.rest.MatrixUserInfo;
 import org.beta.authentification.matrix.exception.MatrixRuntimeException;
 import org.beta.authentification.matrix.rest.MatrixService;
 import org.beta.authentification.matrix.rest.room.DirectRoomsResource;
